Large re-factor: removed Strings in favour of slices
The parser and interpreter now use string slices instead of Strings in
order to reduce unnecessary heap allocations.  Strings in the AST and
during the interpretation stage are now simply references to strings
within the script source rather than String clones.

This change required a lot of explicit lifetime annotations throughout
the code-base.  The 'src lifetime is now used throughout and represents
the lifetime of the script source code being parsed and interpreted.